+++ src/Steps/Refiners/String/StrAfterLast.php
@@ -2,28 +2,22 @@
 
 namespace Crwlr\Crawler\Steps\Refiners\String;
 
-use Crwlr\Crawler\Steps\Refiners\AbstractRefiner;
-
-class StrAfterLast extends AbstractRefiner
+class StrAfterLast extends AbstractStringRefiner
 {
     public function __construct(protected readonly string $last) {}
 
     public function refine(mixed $value): mixed
     {
-        if (!is_string($value)) {
-            $this->logTypeWarning('StringRefiner::afterLast()', $value);
+        return $this->apply($value, function ($value) {
+            if ($this->last === '') {
+                return '';
+            }
 
-            return $value;
-        }
+            $split = explode($this->last, $value);
 
-        if ($this->last === '') {
-            return '';
-        }
+            $lastPart = end($split);
 
-        $split = explode($this->last, $value);
-
-        $lastPart = end($split);
-
-        return trim($lastPart);
+            return trim($lastPart);
+        }, 'StringRefiner::afterLast()');
     }
 }