+++ tests/Steps/Refiners/String/BeforeLastTest.php
@@ -15,14 +15,22 @@ it('logs a warning and returns the unchanged value when $value is not of type st
 
     $logOutput = $this->getActualOutputForAssertion();
 
-    expect($logOutput)->toContain('Refiner StringRefiner::beforeLast() can\'t be applied to value of type ' . gettype($value));
-
-    expect($refinedValue)->toBe($value);
+    expect($logOutput)
+        ->toContain('Refiner StringRefiner::beforeLast() can\'t be applied to value of type ' . gettype($value))
+        ->and($refinedValue)->toBe($value);
 })->with([
     [123],
     [12.3],
     [true],
 ]);
+
+it('works with an array of strings as value', function () {
+    $refinedValue = StringRefiner::beforeLast('a')
+        ->addLogger(new CliLogger())
+        ->refine(['foo a bar a baz', 'lorem a ipsum a dolor']);
+
+    expect($refinedValue)->toBe(['foo a bar a b', 'lorem a ipsum']);
+});
 
 it('returns the string before the last occurrence of another string', function () {
     expect(StringRefiner::beforeLast('foo')->refine('yo lo foo boo choo foo gnu'))->toBe('yo lo foo boo choo');