+++ tests/Steps/Refiners/String/ReplaceTest.php
@@ -15,14 +15,22 @@ it('logs a warning and returns the unchanged value when $value is not of type st
 
     $logOutput = $this->getActualOutputForAssertion();
 
-    expect($logOutput)->toContain('Refiner StringRefiner::replace() can\'t be applied to value of type ' . gettype($value));
-
-    expect($refinedValue)->toBe($value);
+    expect($logOutput)
+        ->toContain('Refiner StringRefiner::replace() can\'t be applied to value of type ' . gettype($value))
+        ->and($refinedValue)->toBe($value);
 })->with([
     [123],
     [12.3],
     [true],
 ]);
+
+it('works when the value is an array of strings', function () {
+    $refinedValue = StringRefiner::replace('foo', 'bar')
+        ->addLogger(new CliLogger())
+        ->refine(['foo boo', 'who foo', 'yo lo']);
+
+    expect($refinedValue)->toBe(['bar boo', 'who bar', 'yo lo']);
+});
 
 it('replaces occurrences of a string with another string', function () {
     expect(StringRefiner::replace('foo', 'bar')->refine('foo, test lorem foo yolo'))->toBe('bar, test lorem bar yolo');