+++ tests/Steps/Refiners/Url/WithPortTest.php
@@ -37,3 +37,13 @@ it('replaces the port in a URL', function (mixed $value, string $expected) {
     [Url::parse('https://www.crwlr.software:5678/bar'), 'https://www.crwlr.software:1234/bar'],
     [Url::parsePsr7('https://crwl.io/quz'), 'https://crwl.io:1234/quz'],
 ]);
+
+it('refines an array of URLs', function () {
+    expect(
+        UrlRefiner::withPort(1234)
+            ->refine([
+                'https://www.example.com/foo',
+                'https://www.example.com/bar',
+            ])
+    )->toBe(['https://www.example.com:1234/foo', 'https://www.example.com:1234/bar']);
+});